Setting the plotting class outside to easy management and calls
